indexmanager: moved files
