+++ app/render/CMakeLists.txt
@@ -26,8 +26,6 @@ set(OLIVE_SOURCES
   render/colorprocessor.cpp
   render/diskmanager.h
   render/diskmanager.cpp
-  render/indexmanager.h
-  render/indexmanager.cpp
   render/pixelformat.h
   render/pixelformat.cpp
   render/pixelservice.h
