+++ app/render/backend/CMakeLists.txt
@@ -32,6 +32,8 @@ set(OLIVE_SOURCES
   render/backend/audiorenderbackend.cpp
   render/backend/audiorenderworker.h
   render/backend/audiorenderworker.cpp
+  render/backend/indexmanager.h
+  render/backend/indexmanager.cpp
   
   render/backend/videorenderbackend.h
   render/backend/videorenderbackend.cpp
