+++ app/render/backend/indexmanager.cpp
