+++ app/render/backend/indexmanager.h
